requirements.txt
streamlit==1.44.1
python-dotenv==1.0.1
openai==1.75.0
pytest==8.3.5
